requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cffi==1.17.0
cryptography==44.0.0
distlib==0.3.9
Django==5.1.5
fastapi==0.115.12
filelock==3.17.0
idna==3.7
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
redis==5.2.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.1
typing_extensions==4.13.0
tzdata==2024.2
uvicorn==0.34.0
virtualenv==20.29.1

